(chore) improve edit and add content of units
